requirements.txt
tweepy>=4.0.0,<5.0.0
requests
python-dotenv

